confirm delete

+++ src/components/HistoryOperations/AllOperations/AllOperations.tsx
@@ -198,20 +198,23 @@ const HistoryOperations: React.FC = () => {
   );
 
   const handleDelete = async (id: any, type: string) => {
-    setDeletingOperation(id);
-    dispatchTyped(
-      type ? deleteOperation(id) : deleteTransferOperation(id)
-    ).then(() => {
-      setDeletingOperation(null);
-      dispatchTyped(getAllOperations());
-      const updatedList = filteredOperations.filter(
-        (operation) => operation._id !== id
-      );
-      const totalPages = Math.ceil(updatedList.length / ITEMS_PER_PAGE);
-      if (currentPage > totalPages) {
-        setCurrentPage(totalPages);
-      }
-    });
+    const shouldDelete = window.confirm("Видалити запис?");
+    if (shouldDelete) {
+      setDeletingOperation(id);
+      dispatchTyped(
+        type ? deleteOperation(id) : deleteTransferOperation(id)
+      ).then(() => {
+        setDeletingOperation(null);
+        dispatchTyped(getAllOperations());
+        const updatedList = filteredOperations.filter(
+          (operation) => operation._id !== id
+        );
+        const totalPages = Math.ceil(updatedList.length / ITEMS_PER_PAGE);
+        if (currentPage > totalPages) {
+          setCurrentPage(totalPages);
+        }
+      });
+    }
   };
 
   return (
